Move isValidFilePath(..) into a utility class.
